Revert "gralloc: Add support for CRC calculation"
This commit is from the latest merge.
I would like to include CRC calculation at some point
because it should improve GPU performance,
but it currently breaks compile because
the Mapper Extension 1.3 HAL is completely
different in T vs R (with GRF, 8350 is still
built in R). Reverting for now, but will
revisit.

This reverts commit 10d758df9d3ff2fab47d0f5dd1272f6cf6f7fbdd.

Change-Id: Ib205c389754b126626dc73c7680b3e2fae780b11